requirements.txt
pyspark==3.5.0
python-dotenv==1.0.0
pandas
numpy==1.26.4
psycopg2-binary==2.9.9
boto3==1.34.0
requests==2.31.0
paramiko==3.4.0
openpyxl==3.1.2
delta-spark==3.0.0